Menuを修正

+++ Assets/Scripts/ToffMonaka/Lib/Scene/MainSceneScript.cs
@@ -36,7 +36,7 @@ public abstract class MainSceneScript : ToffMonaka.Lib.Scene.Script
      */
     public MainSceneScript()
     {
-        this._SetScriptType((int)ToffMonaka.Lib.Constant.Util.SCENE.SCRIPT_TYPE.MAIN_SCENE);
+        this._SetScriptType(ToffMonaka.Lib.Constant.Util.SCENE.SCRIPT_TYPE.MAIN_SCENE);
         this._SetActiveAutoFlag(false);
 
         return;
