+++ Assets/Scripts/ToffMonaka/Lib/Scene/ObjectScript.cs
@@ -27,7 +27,7 @@ public abstract class ObjectScript : ToffMonaka.Lib.Scene.Script
      */
     public ObjectScript()
     {
-        this._SetScriptType((int)ToffMonaka.Lib.Constant.Util.SCENE.SCRIPT_TYPE.OBJECT);
+        this._SetScriptType(ToffMonaka.Lib.Constant.Util.SCENE.SCRIPT_TYPE.OBJECT);
         this._SetActiveAutoFlag(true);
 
         return;
